ARWArwing, WM_LevelControl: the local xyz triples are Vec3f
Both files already had the canonical Vec3f in scope -- ARWArwing.c uses it
two lines below arwarwing_setVelocity, and the sky-lighting table in
WM_LevelCon.c is a plain direction-vector array. Retire the two private
copies, and give arwarwing_setVelocity the same const Vec3f* parameter its
arwarwing_addVelocity sibling already has, dropping the nine (int)& casts at
the call sites in 667, 700_Andross and 701.

All five objects byte-identical apart from one anonymous @N pool-symbol
renumber in ARWArwing.o (section contents and relocation offsets unchanged);
main.dol OK.

+++ src/dlls/objects/667/667.c
@@ -187,7 +187,7 @@ void arwingandrossstuff_hitDetect(GameObject* obj)
             v.y = -10.0f * mathCosf(ang);
             v.z = 0.0f;
             w = v;
-            arwarwing_setVelocity(arwing, (int)&w);
+            arwarwing_setVelocity(arwing, &w);
             doRumble(5.0f);
         }
         if (((ObjHitsPriorityState*)objAnim->hitReactState)->lastHitObject == (u32)arwing)
